+++ src/main/java/net/frozenblock/wilderwild/registry/WWSurfaceRules.java
@@ -359,9 +359,9 @@ public final class WWSurfaceRules implements SurfaceRuleEvents.OverworldSurfaceR
 	}
 
 	@NotNull
-	public static SurfaceRules.RuleSource autumnalPlainsRules() {
+	public static SurfaceRules.RuleSource tundraRules() {
 		return SurfaceRules.ifTrue(
-			SurfaceRules.isBiome(WWBiomes.AUTUMNAL_PLAINS),
+			SurfaceRules.isBiome(WWBiomes.TUNDRA),
 			SurfaceRules.ifTrue(SurfaceRules.ON_FLOOR,
 				SurfaceRules.ifTrue(
 					SurfaceRules.waterBlockCheck(-1, 0),
@@ -479,7 +479,7 @@ public final class WWSurfaceRules implements SurfaceRuleEvents.OverworldSurfaceR
 				rainforestRules(),
 				dyingForestRules(),
 				mapleForestRules(),
-				autumnalPlainsRules()
+				tundraRules()
 			)
 		);
 		WWConstants.log("Wilder Wild's Overworld Surface Rules have been added!", true);